Refactored and organized log view into readable components.
