+++ package.json
@@ -50,7 +50,6 @@
     "onCommand:vscode-docker.create-ACR-Registry",
     "onCommand:vscode-docker.system.prune",
     "onCommand:vscode-docker.dockerHubLogout",
-    "onCommand:vscode-docker.queuebuild",
     "onCommand:vscode-docker.browseDockerHub",
     "onCommand:vscode-docker.browseAzurePortal",
     "onCommand:vscode-docker.explorer.refresh",
@@ -64,7 +63,8 @@
   "main": "./out/dockerExtension",
   "contributes": {
     "menus": {
-      "commandPalette": [{
+      "commandPalette": [
+        {
           "command": "vscode-docker.browseDockerHub",
           "when": "false"
         },
@@ -73,14 +73,10 @@
           "when": "false"
         }
       ],
-      "editor/context": [{
-          "when": "editorLangId == dockerfile",
-          "command": "vscode-docker.image.build",
-          "group": "docker"
-        },
+      "editor/context": [
         {
           "when": "editorLangId == dockerfile",
-          "command": "vscode-docker.queueBuild",
+          "command": "vscode-docker.image.build",
           "group": "docker"
         },
         {
@@ -114,14 +110,10 @@
           "group": "docker"
         }
       ],
-      "explorer/context": [{
-          "when": "resourceFilename =~ /[dD]ocker[fF]ile/",
-          "command": "vscode-docker.image.build",
-          "group": "docker"
-        },
+      "explorer/context": [
         {
           "when": "resourceFilename =~ /[dD]ocker[fF]ile/",
-          "command": "vscode-docker.queueBuild",
+          "command": "vscode-docker.image.build",
           "group": "docker"
         },
         {
@@ -140,7 +132,8 @@
           "group": "docker"
         }
       ],
-      "view/title": [{
+      "view/title": [
+        {
           "command": "vscode-docker.explorer.refresh",
           "when": "view == dockerExplorer",
           "group": "navigation"
@@ -151,7 +144,8 @@
           "group": "navigation"
         }
       ],
-      "view/item/context": [{
+      "view/item/context": [
+        {
           "command": "vscode-docker.container.start",
           "when": "view == dockerExplorer && viewItem =~ /^(localImageNode|imagesRootNode)$/"
         },
@@ -253,34 +247,40 @@
         }
       ]
     },
-    "debuggers": [{
-      "type": "docker",
-      "label": "Docker",
-      "configurationSnippets": [{
-        "label": "Docker: Attach to Node",
-        "description": "Docker: Attach to Node",
-        "body": {
-          "type": "node",
-          "request": "attach",
-          "name": "Docker: Attach to Node",
-          "port": 9229,
-          "address": "localhost",
-          "localRoot": "^\"\\${workspaceFolder}\"",
-          "remoteRoot": "/usr/src/app",
-          "protocol": "inspector"
-        }
-      }]
-    }],
-    "languages": [{
-      "id": "dockerfile",
-      "aliases": [
-        "Dockerfile"
-      ],
-      "filenamePatterns": [
-        "*.dockerfile",
-        "Dockerfile"
-      ]
-    }],
+    "debuggers": [
+      {
+        "type": "docker",
+        "label": "Docker",
+        "configurationSnippets": [
+          {
+            "label": "Docker: Attach to Node",
+            "description": "Docker: Attach to Node",
+            "body": {
+              "type": "node",
+              "request": "attach",
+              "name": "Docker: Attach to Node",
+              "port": 9229,
+              "address": "localhost",
+              "localRoot": "^\"\\${workspaceFolder}\"",
+              "remoteRoot": "/usr/src/app",
+              "protocol": "inspector"
+            }
+          }
+        ]
+      }
+    ],
+    "languages": [
+      {
+        "id": "dockerfile",
+        "aliases": [
+          "Dockerfile"
+        ],
+        "filenamePatterns": [
+          "*.dockerfile",
+          "Dockerfile"
+        ]
+      }
+    ],
     "configuration": {
       "type": "object",
       "title": "Docker configuration options",
@@ -440,7 +440,8 @@
         }
       }
     },
-    "commands": [{
+    "commands": [
+      {
         "command": "vscode-docker.configure",
         "title": "Add Docker files to Workspace",
         "description": "Add Dockerfile, docker-compose.yml files",
@@ -611,29 +612,28 @@
         "category": "Docker"
       },
       {
-        "command": "vscode-docker.queueBuild",
-        "title": "Cloud Build",
-        "category": "Docker"
-      },
-      {
         "command": "vscode-docker.acrBuildLogs",
         "title": "ACR build logs",
         "category": "Docker"
       }
     ],
     "views": {
-      "dockerView": [{
-        "id": "dockerExplorer",
-        "name": "Explorer",
-        "when": "config.docker.showExplorer == true"
-      }]
+      "dockerView": [
+        {
+          "id": "dockerExplorer",
+          "name": "Explorer",
+          "when": "config.docker.showExplorer == true"
+        }
+      ]
     },
     "viewsContainers": {
-      "activitybar": [{
-        "icon": "images/docker.svg",
-        "id": "dockerView",
-        "title": "Docker"
-      }]
+      "activitybar": [
+        {
+          "icon": "images/docker.svg",
+          "id": "dockerView",
+          "title": "Docker"
+        }
+      ]
     }
   },
   "engines": {
@@ -680,6 +680,7 @@
     "fs-extra": "^6.0.1",
     "glob": "7.1.2",
     "gradle-to-js": "^1.0.1",
+    "handlebars": "^4.0.11",
     "moment": "^2.19.3",
     "opn": "^5.1.0",
     "pom-parser": "^1.1.1",
